chore: separate directory
