+++ src/components/Input/CurrencyInput/index.ts
@@ -1,0 +1,1 @@
+export { CurrencyInput } from './CurrencyInput'
